simplify managing crossteams
